general: refactored ContentType to derive EnumOption to allow re-use of FeatureOptions.CreateFeatureOptionsSelectionsView().. and converted cleaner pinball content types as the first consumer
