+++ ClrVpin/Cleaner/CleanerResultsViewModel.cs
@@ -56,7 +56,7 @@ namespace ClrVpin.Cleaner
                 SelectedCommand = new ActionCommand(UpdateHitsView)
             }).ToList();
 
-            return featureTypes.Concat(new[] { FeatureType.CreateSelectAll(featureTypes) }).ToList();
+            return featureTypes.Concat(new[] { FeatureOptions.CreateSelectAll(featureTypes) }).ToList();
         }
 
         protected override IList<FeatureType> CreateAllHitFeatureTypes()
@@ -71,7 +71,7 @@ namespace ClrVpin.Cleaner
                 SelectedCommand = new ActionCommand(UpdateHitsView)
             }).ToList();
 
-            return featureTypes.Concat(new[] { FeatureType.CreateSelectAll(featureTypes) }).ToList();
+            return featureTypes.Concat(new[] { FeatureOptions.CreateSelectAll(featureTypes) }).ToList();
         }
