+++ ClrVpin/Models/Shared/ContentType.cs
@@ -6,14 +6,12 @@ using Utils.Extensions;
 namespace ClrVpin.Models.Shared
 {
     [AddINotifyPropertyChangedInterface]
-    public class ContentType : FolderType
+    public class ContentType : FolderType<ContentTypeEnum>
     {
-        public ContentTypeEnum Enum { get; set; }
         public ContentTypeCategoryEnum Category { get; set; }
 
         public string Extensions { get; set; }
         public string KindredExtensions { get; set; }
-        public string Tip { get; set; }
 
         public IEnumerable<string> ExtensionsList => Extensions?.Split(",").Select(x => x.Trim()) ?? new List<string>();
         public IEnumerable<string> KindredExtensionsList => KindredExtensions?.Split(",").Select(x => x.Trim()) ?? new List<string>();
