+++ ClrVpin/Models/Shared/FolderType.cs
@@ -1,10 +1,11 @@
-﻿using PropertyChanged;
+﻿using System;
+using ClrVpin.Models.Feeder;
+using PropertyChanged;
 
 namespace ClrVpin.Models.Shared;
 
 [AddINotifyPropertyChangedInterface]
-public class FolderType
+public class FolderType<T> : EnumOption<T> where T : Enum
 {
     public string Folder { get; set; }
-    public string Description { get; set; }
 }
